Module re_sdk::components

These are the different components you can log.

They all implement the Component trait, and can be used in MsgSender::with_component.

Structs

AnnotationContext

The AnnotationContext provides additional information on how to display entities.

Box3D

A 3D bounding box represented by it’s half-lengths

ClassId

A 16-bit ID representing a type of semantic class.

Color

An RGBA color tuple with unmultiplied/separate alpha, in sRGB gamma space with linear alpha.

DisconnectedSpace

Specifies that the entity path at which this is logged is disconnected from its parent.

DrawOrder

Draw order used for the display order of 2D elements.

EncodedMesh3D

Compressed/encoded mesh format

InstanceKey

A unique numeric identifier for each individual instance within a batch.

KeypointId

A 16-bit ID representing a type of semantic keypoint within a class.

Label

A String label component.

LineStrip2D

A line strip in 2D space.

LineStrip3D

A line strip in 3D space.

MeshId

A unique id per Mesh3D.

Origin3D

A point of origin in 3D space.

Pinhole

Camera perspective projection (a.k.a. intrinsics).

Point2D

A point in 2D space.

Point3D

A point in 3D space.

Quaternion

A Quaternion represented by 4 real numbers.

Radius

A Radius component.

RawMesh3D

A raw “triangle soup” mesh.

Scalar

A double-precision scalar.

ScalarPlotProps

Additional properties of a scalar when rendered as a plot.

Tensor

A Multi-dimensional Tensor.

TensorDimension

Flattened Tensor data payload

TensorId

A unique id per Tensor.

TextEntry

A text entry component, comprised of a text body and its log level.

Transform3D

An affine transform between two 3D spaces, represented in a given direction.

Vector3D

A vector in 3D space.

ViewCoordinates

How we interpret the coordinate system of an entity/space.

Enums

Mesh3D

A Generic 3D Mesh.

MeshFormat

The format of a binary mesh file, e.g. GLTF, GLB, OBJ

Rect2D

A rectangle in 2D space.

TensorData

Flattened Tensor data payload

TensorDataMeaning

How to interpret the contents of a tensor.