re_sdk
Module external

Re-exports of other crates.

Re-exports

pub use re_grpc_client;

pub use re_grpc_server;

pub use re_log;

pub use re_log_encoding;

pub use re_log_types;

pub use re_data_loader;

Modules

arrow

A complete, safe, native Rust implementation of Apache Arrow, a cross-language development platform for in-memory data.

crossbeam

Tools for concurrent programming.

log

A lightweight logging facade.

nohash_hasher

re_byte_size

Calculate the heap-allocated size of values at runtime.

re_log_types

The different types that make up the rerun log format.

re_tuid

TUID: Time-based Unique Identifiers.

re_types_core

The core types and traits that power Rerun’s data model.