re_sdk::external::arrow::array
Struct ArrayDataBuilder

pub struct ArrayDataBuilder {
    data_type: DataType,
    len: usize,
    null_count: Option<usize>,
    null_bit_buffer: Option<Buffer>,
    nulls: Option<NullBuffer>,
    offset: usize,
    buffers: Vec<Buffer>,
    child_data: Vec<ArrayData>,
    align_buffers: bool,
    skip_validation: UnsafeFlag,
}

Builder for ArrayData type

Fields
data_type: DataTypelen: usizenull_count: Option<usize>null_bit_buffer: Option<Buffer>nulls: Option<NullBuffer>offset: usizebuffers: Vec<Buffer>child_data: Vec<ArrayData>align_buffers: boolskip_validation: UnsafeFlag
Implementations

impl ArrayDataBuilder

pub const fn new(data_type: DataType) -> ArrayDataBuilder

Creates a new array data builder

pub fn data_type(self, data_type: DataType) -> ArrayDataBuilder

Creates a new array data builder from an existing one, changing the data type

pub const fn len(self, n: usize) -> ArrayDataBuilder

Sets the length of the ArrayData

pub fn nulls(self, nulls: Option<NullBuffer>) -> ArrayDataBuilder

Sets the null buffer of the ArrayData

pub fn null_count(self, null_count: usize) -> ArrayDataBuilder

Sets the null count of the ArrayData

pub fn null_bit_buffer(self, buf: Option<Buffer>) -> ArrayDataBuilder

Sets the null_bit_buffer of the ArrayData

pub const fn offset(self, n: usize) -> ArrayDataBuilder

Sets the offset of the ArrayData

pub fn buffers(self, v: Vec<Buffer>) -> ArrayDataBuilder

Sets the buffers of the ArrayData

pub fn add_buffer(self, b: Buffer) -> ArrayDataBuilder

Adds a single buffer to the ArrayDataโs buffers

pub fn add_buffers<I>(self, bs: I) -> ArrayDataBuilder
where I: IntoIterator<Item = Buffer>,

Adds multiple buffers to the ArrayDataโs buffers

pub fn child_data(self, v: Vec<ArrayData>) -> ArrayDataBuilder

Sets the child data of the ArrayData

pub fn add_child_data(self, r: ArrayData) -> ArrayDataBuilder

Adds a single child data to the ArrayDataโs child data

pub unsafe fn build_unchecked(self) -> ArrayData

Creates an array data, without any validation

Note: This is shorthand for

builder.skip_validation(true).build().unwrap()

Safety

The same caveats as ArrayData::new_unchecked apply.

pub fn build(self) -> Result<ArrayData, ArrowError>

Creates an ArrayData, consuming self

Safety

By default the underlying buffers are checked to ensure they are valid Arrow data. However, if the Self::skip_validation flag has been set to true (by the unsafe API) this validation is skipped. If the data is not valid, undefined behavior will result.

pub fn build_aligned(self) -> Result<ArrayData, ArrowError>

๐Deprecated since 54.1.0: Use ArrayData::align_buffers instead

Creates an array data, validating all inputs, and aligning any buffers

pub fn align_buffers(self, align_buffers: bool) -> ArrayDataBuilder

Ensure that all buffers are aligned, copying data if necessary

Rust requires that arrays are aligned to their corresponding primitive, see Layout::array and std::mem::align_of.

ArrayData therefore requires that all buffers have at least this alignment, to allow for slice based APIs. See BufferSpec::FixedWidth.

As this alignment is architecture specific, and not guaranteed by all arrow implementations, this flag is provided to automatically copy buffers to a new correctly aligned allocation when necessary, making it useful when interacting with buffers produced by other systems, e.g. IPC or FFI.

If this flag is not enabled, [Self::build] return an error on encountering insufficiently aligned buffers.

pub unsafe fn skip_validation(self, skip_validation: bool) -> ArrayDataBuilder

Skips validation of the data.

If this flag is enabled, [Self::build] will skip validation of the data

If this flag is not enabled, [Self::build] will validate that all buffers are valid and will return an error if any data is invalid. Validation can be expensive.

Safety

If validation is skipped, the buffers must form a valid Arrow array, otherwise undefined behavior will result