Module re_sdk::external::arrow::array::types

Zero-sized types used to parameterize generic array implementations

Structs

BinaryViewType

BinaryViewType for string arrays

BooleanType

A boolean datatype

Date32Type

32-bit date type: the elapsed time since UNIX epoch in days (32 bits).

Date64Type

64-bit date type: the elapsed time since UNIX epoch in milliseconds (64 bits). Values must be divisible by 86_400_000. See DataType::Date64 for more details.

Decimal128Type

The decimal type for a Decimal128Array

Decimal256Type

The decimal type for a Decimal256Array

DurationMicrosecondType

Elapsed time type: microseconds.

DurationMillisecondType

Elapsed time type: milliseconds.

DurationNanosecondType

Elapsed time type: nanoseconds.

DurationSecondType

Elapsed time type: seconds.

Float16Type

16-bit floating point number type.

Float32Type

32-bit floating point number type.

Float64Type

64-bit floating point number type.

GenericBinaryType

ByteArrayType for binary arrays

GenericStringType

ByteArrayType for string arrays

Int8Type

A signed 8-bit integer type.

Int16Type

Signed 16-bit integer type.

Int32Type

Signed 32-bit integer type.

Int64Type

Signed 64-bit integer type.

IntervalDayTime

Value of an IntervalDayTime array

IntervalDayTimeType

“Calendar” interval type: days and milliseconds. See IntervalDayTime for more details.

IntervalMonthDayNano

Value of an IntervalMonthDayNano array

IntervalMonthDayNanoType

“Calendar” interval type: months, days, and nanoseconds. See IntervalMonthDayNano for more details.

IntervalYearMonthType

32-bit “calendar” interval type: the number of whole months.

StringViewType

ByteViewType for string arrays

Time32MillisecondType

32-bit time type: the elapsed time since midnight in milliseconds.

Time32SecondType

32-bit time type: the elapsed time since midnight in seconds.

Time64MicrosecondType

64-bit time type: the elapsed time since midnight in microseconds.

Time64NanosecondType

64-bit time type: the elapsed time since midnight in nanoseconds.

TimestampMicrosecondType

Timestamp microsecond type with an optional timezone.

TimestampMillisecondType

Timestamp millisecond type with an optional timezone.

TimestampNanosecondType

Timestamp nanosecond type with an optional timezone.

TimestampSecondType

Timestamp second type with an optional timezone.

UInt8Type

Unsigned 8-bit integer type.

UInt16Type

Unsigned 16-bit integer type.

UInt32Type

Unsigned 32-bit integer type.

UInt64Type

Unsigned 64-bit integer type.

Traits

ArrowDictionaryKeyType

A subtype of primitive type that represents legal dictionary keys. See https://arrow.apache.org/docs/format/Columnar.html

ArrowPrimitiveType

Trait for primitive values.

ArrowTemporalType

A subtype of primitive type that represents temporal values.

ArrowTimestampType

A timestamp type allows us to create array builders that take a timestamp.

ByteArrayType

A trait over the variable-size byte array types

ByteViewType

A trait over the variable length bytes view array types

DecimalType

A trait over the decimal types, used by PrimitiveArray to provide a generic implementation across the various decimal types

RunEndIndexType

A subtype of primitive type that is used as run-ends index in RunArray. See https://arrow.apache.org/docs/format/Columnar.html

Functions

validate_decimal_precision_and_scale

Validate that precision and scale are valid for T

Type Aliases

BinaryType

An arrow binary array with i32 offsets

LargeBinaryType

An arrow binary array with i64 offsets

LargeUtf8Type

An arrow utf8 array with i64 offsets

Utf8Type

An arrow utf8 array with i32 offsets