Module re_sdk::external::arrow::buffer

Types of shared memory region

Structs

BooleanBuffer

A slice-able Buffer containing bit-packed booleans

Buffer

Buffer represents a contiguous memory region that can be shared with other buffers and across thread boundaries.

MutableBuffer

A MutableBuffer is Arrow’s interface to build a Buffer out of items or slices of items.

NullBuffer

A BooleanBuffer used to encode validity for arrow arrays

OffsetBuffer

A non-empty buffer of monotonically increasing, positive integers.

RunEndBuffer

A slice-able buffer of monotonically increasing, positive integers used to store run-ends

ScalarBuffer

A strongly-typed Buffer supporting zero-copy cloning and slicing

Functions

bitwise_bin_op_helper

Apply a bitwise operation op to two inputs and return the result as a Buffer. The inputs are treated as bitmaps, meaning that offsets and length are specified in number of bits.

bitwise_quaternary_op_helper

Apply a bitwise operation op to four inputs and return the result as a Buffer. The inputs are treated as bitmaps, meaning that offsets and length are specified in number of bits.

bitwise_unary_op_helper

Apply a bitwise operation op to one input and return the result as a Buffer. The input is treated as a bitmap, meaning that offset and length are specified in number of bits.

buffer_bin_and

Apply a bitwise and to two inputs and return the result as a Buffer. The inputs are treated as bitmaps, meaning that offsets and length are specified in number of bits.

buffer_bin_and_not

Apply a bitwise and_not to two inputs and return the result as a Buffer. The inputs are treated as bitmaps, meaning that offsets and length are specified in number of bits.

buffer_bin_or

Apply a bitwise or to two inputs and return the result as a Buffer. The inputs are treated as bitmaps, meaning that offsets and length are specified in number of bits.

buffer_bin_xor

Apply a bitwise xor to two inputs and return the result as a Buffer. The inputs are treated as bitmaps, meaning that offsets and length are specified in number of bits.

buffer_unary_not

Apply a bitwise not to one input and return the result as a Buffer. The input is treated as a bitmap, meaning that offset and length are specified in number of bits.