re_sdk::external::arrow::datatypes
Struct Schema

pub struct Schema {
    pub fields: Fields,
    pub metadata: HashMap<String, String>,
}

Describes the meta-data of an ordered sequence of relative types.

Note that this information is only part of the meta-data and not part of the physical memory layout.

Fields
fields: Fields

A sequence of fields that describe the schema.

metadata: HashMap<String, String>

A map of key-value pairs containing additional meta data.

Implementations

impl Schema

pub fn empty() -> Schema

Creates an empty Schema

pub fn new(fields: impl Into<Fields>) -> Schema

Creates a new Schema from a sequence of Field values.

Example

let field_a = Field::new("a", DataType::Int64, false);
let field_b = Field::new("b", DataType::Boolean, false);

let schema = Schema::new(vec![field_a, field_b]);

pub fn new_with_metadata( fields: impl Into<Fields>, metadata: HashMap<String, String>, ) -> Schema

Creates a new Schema from a sequence of Field values and adds additional metadata in form of key value pairs.

Example

let field_a = Field::new("a", DataType::Int64, false);
let field_b = Field::new("b", DataType::Boolean, false);

let mut metadata: HashMap<String, String> = HashMap::new();
metadata.insert("row_count".to_string(), "100".to_string());

let schema = Schema::new_with_metadata(vec![field_a, field_b], metadata);

pub fn with_metadata(self, metadata: HashMap<String, String>) -> Schema

Sets the metadata of this Schema to be metadata and returns self

pub fn project(&self, indices: &[usize]) -> Result<Schema, ArrowError>

Returns a new schema with only the specified columns in the new schema This carries metadata from the parent schema over as well

pub fn try_merge( schemas: impl IntoIterator<Item = Schema>, ) -> Result<Schema, ArrowError>

Merge schema into self if it is compatible. Struct fields will be merged recursively.

Example:

let merged = Schema::try_merge(vec![
    Schema::new(vec![
        Field::new("c1", DataType::Int64, false),
        Field::new("c2", DataType::Utf8, false),
    ]),
    Schema::new(vec![
        Field::new("c1", DataType::Int64, true),
        Field::new("c2", DataType::Utf8, false),
        Field::new("c3", DataType::Utf8, false),
    ]),
]).unwrap();

assert_eq!(
    merged,
    Schema::new(vec![
        Field::new("c1", DataType::Int64, true),
        Field::new("c2", DataType::Utf8, false),
        Field::new("c3", DataType::Utf8, false),
    ]),
);

pub const fn fields(&self) -> &Fields

Returns an immutable reference of the vector of Field instances.

pub fn flattened_fields(&self) -> Vec<&Field>

Returns a vector with references to all fields (including nested fields)

Example

use std::sync::Arc;
use arrow_schema::{DataType, Field, Fields, Schema};

let f1 = Arc::new(Field::new("a", DataType::Boolean, false));

let f2_inner = Arc::new(Field::new("b_inner", DataType::Int8, false));
let f2 = Arc::new(Field::new("b", DataType::List(f2_inner.clone()), false));

let f3_inner1 = Arc::new(Field::new("c_inner1", DataType::Int8, false));
let f3_inner2 = Arc::new(Field::new("c_inner2", DataType::Int8, false));
let f3 = Arc::new(Field::new(
    "c",
    DataType::Struct(vec![f3_inner1.clone(), f3_inner2.clone()].into()),
    false
));

let mut schema = Schema::new(vec![
  f1.clone(), f2.clone(), f3.clone()
]);
assert_eq!(
    schema.flattened_fields(),
    vec![
        f1.as_ref(),
        f2.as_ref(),
        f2_inner.as_ref(),
        f3.as_ref(),
        f3_inner1.as_ref(),
        f3_inner2.as_ref()
   ]
);

pub fn all_fields(&self) -> Vec<&Field>

๐Deprecated since 52.2.0: Use flattened_fields instead

Returns a vector with references to all fields (including nested fields)

pub fn field(&self, i: usize) -> &Field

Returns an immutable reference of a specific Field instance selected using an offset within the internal fields vector.

Panics

Panics if index out of bounds

pub fn field_with_name(&self, name: &str) -> Result<&Field, ArrowError>

Returns an immutable reference of a specific Field instance selected by name.

pub fn fields_with_dict_id(&self, dict_id: i64) -> Vec<&Field>

๐Deprecated since 54.0.0: The ability to preserve dictionary IDs will be removed. With it, all functions related to it.

Returns a vector of immutable references to all Field instances selected by the dictionary ID they use.

pub fn index_of(&self, name: &str) -> Result<usize, ArrowError>

Find the index of the column with the given name.

pub const fn metadata(&self) -> &HashMap<String, String>

Returns an immutable reference to the Map of custom metadata key-value pairs.

pub fn normalize( &self, separator: &str, max_level: Option<usize>, ) -> Result<Schema, ArrowError>

Normalize a Schema into a flat table.

Nested Fields will generate names separated by separator, up to a depth of max_level (unlimited if None).

e.g. given a Schema:

    "foo": StructArray<"bar": Utf8>

A separator of "." would generate a batch with the schema:

    "foo.bar": Utf8

Note that giving a depth of Some(0) to max_level is the same as passing in None; it will be treated as unlimited.

Example

let schema = Schema::new(vec![
    Field::new(
        "a",
        DataType::Struct(Fields::from(vec![
            Arc::new(Field::new("animals", DataType::Utf8, true)),
            Arc::new(Field::new("n_legs", DataType::Int64, true)),
        ])),
        false,
    ),
])
.normalize(".", None)
.expect("valid normalization");
let expected = Schema::new(vec![
    Field::new("a.animals", DataType::Utf8, true),
    Field::new("a.n_legs", DataType::Int64, true),
]);
assert_eq!(schema, expected);

pub fn column_with_name(&self, name: &str) -> Option<(usize, &Field)>

Look up a column by name and return a immutable reference to the column along with its index.

pub fn contains(&self, other: &Schema) -> bool

Check to see if self is a superset of other schema.

In particular returns true if self.metadata is a superset of other.metadata and Fields::contains for self.fields and other.fields

In other words, any record that conforms to other should also conform to self.