re_sdk::external::arrow::datatypes
Struct i256

#[repr(C)]
pub struct i256 {
    low: u128,
    high: i128,
}

A signed 256-bit integer

Fields
low: u128high: i128
Implementations

impl i256

pub const ZERO: i256

The additive identity for this integer type, i.e. 0.

pub const ONE: i256

The multiplicative identity for this integer type, i.e. 1.

pub const MINUS_ONE: i256

The multiplicative inverse for this integer type, i.e. -1.

pub const MAX: i256

The maximum value that can be represented by this integer type

pub const MIN: i256

The minimum value that can be represented by this integer type

pub const fn from_le_bytes(b: [u8; 32]) -> i256

Create an integer value from its representation as a byte array in little-endian.

pub const fn from_be_bytes(b: [u8; 32]) -> i256

Create an integer value from its representation as a byte array in big-endian.

pub const fn from_i128(v: i128) -> i256

Create an i256 value from a 128-bit value.

pub fn from_string(value_str: &str) -> Option<i256>

Create an integer value from its representation as string.

pub fn from_f64(v: f64) -> Option<i256>

Create an optional i256 from the provided f64. Returning None if overflow occurred

pub const fn from_parts(low: u128, high: i128) -> i256

Create an i256 from the provided low u128 and high i128

pub const fn to_parts(self) -> (u128, i128)

Returns this i256 as a low u128 and high i128

pub fn to_i128(self) -> Option<i128>

Converts this i256 into an i128 returning None if this would result in truncation/overflow

pub fn as_i128(self) -> i128

Wraps this i256 into an i128

pub const fn to_le_bytes(self) -> [u8; 32]

Return the memory representation of this integer as a byte array in little-endian byte order.

pub const fn to_be_bytes(self) -> [u8; 32]

Return the memory representation of this integer as a byte array in big-endian byte order.

pub fn wrapping_abs(self) -> i256

Computes the absolute value of this i256

pub fn checked_abs(self) -> Option<i256>

Computes the absolute value of this i256 returning None if Self == Self::MIN

pub fn wrapping_neg(self) -> i256

Negates this i256

pub fn checked_neg(self) -> Option<i256>

Negates this i256 returning None if Self == Self::MIN

pub fn wrapping_add(self, other: i256) -> i256

Performs wrapping addition

pub fn checked_add(self, other: i256) -> Option<i256>

Performs checked addition

pub fn wrapping_sub(self, other: i256) -> i256

Performs wrapping subtraction

pub fn checked_sub(self, other: i256) -> Option<i256>

Performs checked subtraction

pub fn wrapping_mul(self, other: i256) -> i256

Performs wrapping multiplication

pub fn checked_mul(self, other: i256) -> Option<i256>

Performs checked multiplication

pub fn wrapping_div(self, other: i256) -> i256

Performs wrapping division

pub fn checked_div(self, other: i256) -> Option<i256>

Performs checked division

pub fn wrapping_rem(self, other: i256) -> i256

Performs wrapping remainder

pub fn checked_rem(self, other: i256) -> Option<i256>

Performs checked remainder

pub fn checked_pow(self, exp: u32) -> Option<i256>

Performs checked exponentiation

pub fn wrapping_pow(self, exp: u32) -> i256

Performs wrapping exponentiation

pub const fn signum(self) -> i256

Returns a number i256 representing sign of this i256.

0 if the number is zero 1 if the number is positive -1 if the number is negative

pub const fn is_negative(self) -> bool

Returns true if this i256 is negative

pub const fn is_positive(self) -> bool

Returns true if this i256 is positive