Trait re_sdk::external::arrow2::array::Array

pub trait Array: Send + Sync + DynClone + 'static {

    // Required methods
    fn as_any(&self) -> &(dyn Any + 'static);
    fn as_any_mut(&mut self) -> &mut (dyn Any + 'static);
    fn len(&self) -> usize;
    fn data_type(&self) -> &DataType;
    fn validity(&self) -> Option<&Bitmap>;
    fn slice(&mut self, offset: usize, length: usize);
    unsafe fn slice_unchecked(&mut self, offset: usize, length: usize);
    fn with_validity(&self, validity: Option<Bitmap>) -> Box<dyn Array>;
    fn to_boxed(&self) -> Box<dyn Array>;

    // Provided methods
    fn is_empty(&self) -> bool { ... }
    fn null_count(&self) -> usize { ... }
    fn is_null(&self, i: usize) -> bool { ... }
    unsafe fn is_null_unchecked(&self, i: usize) -> bool { ... }
    fn is_valid(&self, i: usize) -> bool { ... }
    fn sliced(&self, offset: usize, length: usize) -> Box<dyn Array> { ... }
    unsafe fn sliced_unchecked(
        &self,
        offset: usize,
        length: usize
    ) -> Box<dyn Array> { ... }

}

A trait representing an immutable Arrow array. Arrow arrays are trait objects that are infallibly downcasted to concrete types according to the Array::data_type.

Required Methods

fn as_any(&self) -> &(dyn Any + 'static)

Converts itself to a reference of Any, which enables downcasting to concrete types.

fn as_any_mut(&mut self) -> &mut (dyn Any + 'static)

Converts itself to a mutable reference of Any, which enables mutable downcasting to concrete types.

fn len(&self) -> usize

The length of the Array. Every array has a length corresponding to the number of elements (slots).

fn data_type(&self) -> &DataType

The DataType of the Array. In combination with Array::as_any, this can be used to downcast trait objects (dyn Array) to concrete arrays.

fn validity(&self) -> Option<&Bitmap>

The validity of the Array: every array has an optional Bitmap that, when available specifies whether the array slot is valid or not (null). When the validity is None, all slots are valid.

fn slice(&mut self, offset: usize, length: usize)

Slices this Array.

Implementation

This operation is O(1) over len.

Panic

This function panics iff offset + length > self.len().

unsafe fn slice_unchecked(&mut self, offset: usize, length: usize)

Slices the Array.

Implementation

This operation is O(1).

Safety

The caller must ensure that offset + length <= self.len()

fn with_validity(&self, validity: Option<Bitmap>) -> Box<dyn Array>

Clones this Array with a new new assigned bitmap.

Panic

This function panics iff validity.len() != self.len().

fn to_boxed(&self) -> Box<dyn Array>

Clone a &dyn Array to an owned Box<dyn Array>.

Provided Methods

fn is_empty(&self) -> bool

whether the array is empty

fn null_count(&self) -> usize

The number of null slots on this Array.

Implementation

This is O(1) since the number of null elements is pre-computed.

fn is_null(&self, i: usize) -> bool

Returns whether slot i is null.

Panic

Panics iff i >= self.len().

unsafe fn is_null_unchecked(&self, i: usize) -> bool

Returns whether slot i is null.

Safety

The caller must ensure i < self.len()

fn is_valid(&self, i: usize) -> bool

Returns whether slot i is valid.

Panic

Panics iff i >= self.len().

fn sliced(&self, offset: usize, length: usize) -> Box<dyn Array>

Returns a slice of this Array.

Implementation

This operation is O(1) over len.

Panic

This function panics iff offset + length > self.len().

unsafe fn sliced_unchecked( &self, offset: usize, length: usize ) -> Box<dyn Array>

Returns a slice of this Array.

Implementation

This operation is O(1) over len, as it amounts to increase two ref counts and moving the struct to the heap.

Safety

The caller must ensure that offset + length <= self.len()