Module re_sdk::external::arrow2::bitmap

contains Bitmap and MutableBitmap, containers of bool.

Modules

utils

General utilities for bitmaps representing items where LSB is the first item.

Structs

Bitmap

An immutable container semantically equivalent to Arc<Vec<bool>> but represented as Arc<Vec<u8>> where each boolean is represented as a single bit.

IntoIter

This crates’ equivalent of std::vec::IntoIter for Bitmap.

MutableBitmap

A container of booleans. MutableBitmap is semantically equivalent to Vec<bool>.

Functions

and

Compute bitwise AND operation

binary

Apply a bitwise operation op to two inputs and return the result as a Bitmap.

binary_assign

Apply a bitwise binary operation to a MutableBitmap.

chunk_iter_to_vec

Creates a Vec<u8> from a TrustedLen of BitChunk.

from_chunk_iter_unchecked⚠

Creates a Vec from an Iterator of BitChunk.

or

Compute bitwise OR operation

quaternary

Apply a bitwise operation op to four inputs and return the result as a Bitmap.

ternary

Apply a bitwise operation op to three inputs and return the result as a Bitmap.

unary

Apply a bitwise operation op to one input and return the result as a Bitmap.

unary_assign

Applies a function to every bit of this MutableBitmap in chunks

xor

Compute bitwise XOR operation