Module re_sdk::external::arrow2_convert::deserialize

Implementation and traits for deserializing from Arrow.

Structs

BufferBinaryArray

Internal ArrowArray helper to iterate over a BinaryArray while exposing Buffer slices

BufferBinaryArrayIter

Iterator for for BufferBinaryArray

Traits

ArrowDeserialize

Implemented by ArrowField that can be deserialized from arrow

TryIntoCollection

Top-level API to deserialize from Arrow

Functions

arrow_array_deserialize_iterator

Return an iterator that deserializes an Array to an element of type T

arrow_array_deserialize_iterator_as_type

Returns a typed iterator to a target type from an arrow2::Array