Trait re_sdk::external::crossbeam::epoch::Pointer

pub trait Pointer<T>
where
    T: Pointable + ?Sized,
{
    // Required methods
    fn into_usize(self) -> usize;
    unsafe fn from_usize(data: usize) -> Self;
}

A trait for either Owned or Shared pointers.

Required Methods

fn into_usize(self) -> usize

Returns the machine representation of the pointer.

unsafe fn from_usize(data: usize) -> Self

Returns a new pointer pointing to the tagged pointer data.

Safety

The given data should have been created by Pointer::into_usize(), and one data should not be converted back by Pointer::from_usize() multiple times.

Object Safety

This trait is not object safe.