Struct re_sdk::external::image::codecs::png::PngDecoder

pub struct PngDecoder<R>where
    R: Read,{
    color_type: ColorType,
    reader: Reader<R>,
}

PNG decoder

Fields
color_type: ColorTypereader: Reader<R>
Implementations

impl<R> PngDecoder<R>where R: Read,

pub fn new(r: R) -> Result<PngDecoder<R>, ImageError>

Creates a new decoder that decodes from the stream r

pub fn with_limits(r: R, limits: Limits) -> Result<PngDecoder<R>, ImageError>

Creates a new decoder that decodes from the stream r with the given limits.

pub fn apng(self) -> ApngDecoder<R>

Turn this into an iterator over the animation frames.

Reading the complete animation requires more memory than reading the data from the IDAT frame–multiple frame buffers need to be reserved at the same time. We further do not support compositing 16-bit colors. In any case this would be lossy as the interface of animation decoders does not support 16-bit colors.

If something is not supported or a limit is violated then the decoding step that requires them will fail and an error will be returned instead of the frame. No further frames will be returned.

pub fn is_apng(&self) -> bool

Returns if the image contains an animation.

Note that the file itself decides if the default image is considered to be part of the animation. When it is not the common interpretation is to use it as a thumbnail.

If a non-animated image is converted into an ApngDecoder then its iterator is empty.