Struct re_sdk::external::image::codecs::png::PngReader

pub struct PngReader<R>where
    R: Read,{
    reader: Reader<R>,
    buffer: Vec<u8, Global>,
    index: usize,
}

Png Reader

This reader will try to read the png one row at a time, however for interlaced png files this is not possible and these are therefore read at once.

Fields
reader: Reader<R>buffer: Vec<u8, Global>index: usize