Trait re_sdk::external::image::EncodableLayout

pub trait EncodableLayout: EncodableLayout {
    // Required method
    fn as_bytes(&self) -> &[u8] ;
}

Types which are safe to treat as an immutable byte slice in a pixel layout for image encoding.

Required Methods

fn as_bytes(&self) -> &[u8]

Get the bytes of this value.