Crate re_sdk::external::re_types_core

The core types and traits that power Rerun’s data model.

The Archetype trait is the core of this crate and is a good starting point to get familiar with the code. An archetype is a logical collection of batches of Components that play well with each other.

Rerun (and the underlying Arrow data framework) is designed to work with large arrays of Components, as opposed to single instances. When multiple instances of a Component are put together in an array, they yield a ComponentBatch: the atomic unit of (de)serialization.

Internally, Components are implemented using many different Loggables.

Feature flags

serde — Enable (de)serialization using serde.

Modules

archetypes

Fundamental Archetypes that are implemented in re_types_core directly for convenience and dependency optimization.

components

Fundamental Components that are implemented in re_types_core directly for convenience and dependency optimization.

datatypes

Fundamental datatypes that are implemented in re_types_core directly for convenience and dependency optimization.

external

macros

reflection

Run-time reflection for reading meta-data about components and archetypes.

Macros

delegate_arrow_tuid

Implements crate::Component for any given type that is a simple wrapper (newtype) around a Tuid.

impl_into_cow

Implements From<Self> and From<'a Self> for Cow<Self>.

static_assert_struct_has_fields

Useful macro for staticlly asserting that a struct contains some specific fields.

Structs

ArchetypeName

The fully-qualified name of an Archetype, e.g. rerun.archetypes.Points3D.

ArrowBuffer

Convenience-wrapper around an [arrow::buffer::ScalarBuffer] that is known to contain a a primitive type.

ArrowString

Convenience-wrapper around an arrow Buffer that is known to contain a UTF-8 encoded string.

ComponentName

The fully-qualified name of a Component, e.g. rerun.components.Position2D.

DatatypeName

The fully-qualified name of a [Datatype], e.g. rerun.datatypes.Vec2D.

GenericIndicatorComponent

A generic indicator component that can be specialized for any Archetype.

NamedIndicatorComponent

An arbitrary named indicator component.

SpaceViewClassIdentifier

The unique name of a space view type.

Enums

DeserializationError

MaybeOwnedComponentBatch

Holds either an owned ComponentBatch that lives on heap, or a reference to one.

SerializationError

Constants

DEFAULT_DISPLAY_DECIMALS

Number of decimals shown for all float display methods.

Traits

Archetype

An archetype is a high-level construct that represents a set of Components that usually play well with each other (i.e. they compose nicely).

ArchetypeReflectionMarker

Indicates that the archetype has reflection data available for it.

AsComponents

Describes the interface for interpreting an object as a bundle of Components.

Component

A Component describes semantic data that can be used by any number of Archetypes.

ComponentBatch

A ComponentBatch represents an array’s worth of Component instances.

Loggable

A Loggable represents a single instance in an array of loggable data.

LoggableBatch

A LoggableBatch represents an array’s worth of Loggable instances, ready to be serialized.

ResultExt

SizeBytes

Approximations of stack and heap size for both internal and external types.

View

Views are the panels shown in the viewer’s viewport and the primary means of inspecting & visualizing previously logged data.

Type Aliases

ComponentNameSet

DeserializationResult

SerializationResult

UnorderedComponentNameSet

_Backtrace