re_sdk::external::re_types_core::external::arrow::array::array
Struct GenericListArray

pub struct GenericListArray<OffsetSize>
where
    OffsetSize: OffsetSizeTrait,
{
    data_type: DataType,
    nulls: Option<NullBuffer>,
    values: Arc<dyn Array>,
    value_offsets: OffsetBuffer<OffsetSize>,
}

An array of variable length lists, similar to JSON arrays (e.g. ["A", "B", "C"]). This struct specifically represents the list layout. Refer to GenericListViewArray for the list-view layout.

Lists are represented using offsets into a values child array. Offsets are stored in two adjacent entries of an OffsetBuffer.

Arrow defines ListArray with i32 offsets and LargeListArray with i64 offsets.

Use GenericListBuilder to construct a GenericListArray.

Representation

A ListArray can represent a list of values of any other supported Arrow type. Each element of the ListArray itself is a list which may be empty, may contain NULL and non-null values, or may itself be NULL.

For example, the ListArray shown in the following diagram stores lists of strings. Note that [] represents an empty (length 0), but non NULL list.

┌─────────────┐
│   [A,B,C]   │
├─────────────┤
│     []      │
├─────────────┤
│    NULL     │
├─────────────┤
│     [D]     │
├─────────────┤
│  [NULL, F]  │
└─────────────┘

The values are stored in a child StringArray and the offsets are stored in an OffsetBuffer as shown in the following diagram. The logical values and offsets are shown on the left, and the actual ListArray encoding on the right.

                                        ┌ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─
                                                                ┌ ─ ─ ─ ─ ─ ─ ┐    │
 ┌─────────────┐  ┌───────┐             │     ┌───┐   ┌───┐       ┌───┐ ┌───┐
 │   [A,B,C]   │  │ (0,3) │                   │ 1 │   │ 0 │     │ │ 1 │ │ A │ │ 0  │
 ├─────────────┤  ├───────┤             │     ├───┤   ├───┤       ├───┤ ├───┤
 │      []     │  │ (3,3) │                   │ 1 │   │ 3 │     │ │ 1 │ │ B │ │ 1  │
 ├─────────────┤  ├───────┤             │     ├───┤   ├───┤       ├───┤ ├───┤
 │    NULL     │  │ (3,4) │                   │ 0 │   │ 3 │     │ │ 1 │ │ C │ │ 2  │
 ├─────────────┤  ├───────┤             │     ├───┤   ├───┤       ├───┤ ├───┤
 │     [D]     │  │ (4,5) │                   │ 1 │   │ 4 │     │ │ ? │ │ ? │ │ 3  │
 ├─────────────┤  ├───────┤             │     ├───┤   ├───┤       ├───┤ ├───┤
 │  [NULL, F]  │  │ (5,7) │                   │ 1 │   │ 5 │     │ │ 1 │ │ D │ │ 4  │
 └─────────────┘  └───────┘             │     └───┘   ├───┤       ├───┤ ├───┤
                                                      │ 7 │     │ │ 0 │ │ ? │ │ 5  │
                                        │  Validity   └───┘       ├───┤ ├───┤
    Logical       Logical                  (nulls)   Offsets    │ │ 1 │ │ F │ │ 6  │
     Values       Offsets               │                         └───┘ └───┘
                                                                │    Values   │    │
                (offsets[i],            │   ListArray               (Array)
               offsets[i+1])                                    └ ─ ─ ─ ─ ─ ─ ┘    │
                                        └ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─

Slicing

Slicing a ListArray creates a new ListArray without copying any data, but this means the Self::values and Self::offsets may have “unused” data

For example, calling slice(1, 3) on the ListArray in the above example would result in the following. Note

Values array is unchanged

Offsets do not start at 0, nor cover all values in the Values array.

                                ┌ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─
                                                        ┌ ─ ─ ─ ─ ─ ─ ┐    │  ╔═══╗
                                │                         ╔═══╗ ╔═══╗         ║   ║  Not used
                                                        │ ║ 1 ║ ║ A ║ │ 0  │  ╚═══╝
 ┌─────────────┐  ┌───────┐     │     ┌───┐   ┌───┐       ╠═══╣ ╠═══╣
 │ [] (empty)  │  │ (3,3) │           │ 1 │   │ 3 │     │ ║ 1 ║ ║ B ║ │ 1  │
 ├─────────────┤  ├───────┤     │     ├───┤   ├───┤       ╠═══╣ ╠═══╣
 │    NULL     │  │ (3,4) │           │ 0 │   │ 3 │     │ ║ 1 ║ ║ C ║ │ 2  │
 ├─────────────┤  ├───────┤     │     ├───┤   ├───┤       ╠───╣ ╠───╣
 │     [D]     │  │ (4,5) │           │ 1 │   │ 4 │     │ │ 0 │ │ ? │ │ 3  │
 └─────────────┘  └───────┘     │     └───┘   ├───┤       ├───┤ ├───┤
                                              │ 5 │     │ │ 1 │ │ D │ │ 4  │
                                │             └───┘       ├───┤ ├───┤
                                                        │ │ 0 │ │ ? │ │ 5  │
                                │  Validity               ╠═══╣ ╠═══╣
    Logical       Logical          (nulls)   Offsets    │ ║ 1 ║ ║ F ║ │ 6  │
     Values       Offsets       │                         ╚═══╝ ╚═══╝
                                                        │    Values   │    │
                (offsets[i],    │   ListArray               (Array)
               offsets[i+1])                            └ ─ ─ ─ ─ ─ ─ ┘    │
                                └ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─

Fields
data_type: DataTypenulls: Option<NullBuffer>values: Arc<dyn Array>value_offsets: OffsetBuffer<OffsetSize>
Implementations

impl<OffsetSize> GenericListArray<OffsetSize>
where OffsetSize: OffsetSizeTrait,

pub const DATA_TYPE_CONSTRUCTOR: fn(_: Arc<Field>) -> DataType = _

The data type constructor of list array. The input is the schema of the child array and the output is the DataType, List or LargeList.

pub fn try_new( field: Arc<Field>, offsets: OffsetBuffer<OffsetSize>, values: Arc<dyn Array>, nulls: Option<NullBuffer>, ) -> Result<GenericListArray<OffsetSize>, ArrowError>

Create a new GenericListArray from the provided parts

Errors

Errors if

offsets.len() - 1 != nulls.len()

offsets.last() > values.len()

!field.is_nullable() && values.is_nullable()

field.data_type() != values.data_type()

pub fn new( field: Arc<Field>, offsets: OffsetBuffer<OffsetSize>, values: Arc<dyn Array>, nulls: Option<NullBuffer>, ) -> GenericListArray<OffsetSize>

Create a new GenericListArray from the provided parts

Panics

Panics if Self::try_new returns an error

pub fn new_null(field: Arc<Field>, len: usize) -> GenericListArray<OffsetSize>

Create a new GenericListArray of length len where all values are null

pub fn into_parts( self, ) -> (Arc<Field>, OffsetBuffer<OffsetSize>, Arc<dyn Array>, Option<NullBuffer>)

Deconstruct this array into its constituent parts

pub fn offsets(&self) -> &OffsetBuffer<OffsetSize>

Returns a reference to the offsets of this list

Unlike Self::value_offsets this returns the OffsetBuffer allowing for zero-copy cloning.

Notes: The offsets may not start at 0 and may not cover all values in Self::values. This can happen when the list array was sliced via Self::slice. See documentation for Self for more details.

pub fn values(&self) -> &Arc<dyn Array>

Returns a reference to the values of this list

Note: The list array may not refer to all values in the values array. For example if the list array was sliced via Self::slice values will still contain values both before and after the slice. See documentation for Self for more details.

pub fn value_type(&self) -> DataType

Returns a clone of the value type of this list.

pub unsafe fn value_unchecked(&self, i: usize) -> Arc<dyn Array>

Returns ith value of this list array.

Safety

Caller must ensure that the index is within the array bounds

pub fn value(&self, i: usize) -> Arc<dyn Array>

Returns ith value of this list array.

pub fn value_offsets(&self) -> &[OffsetSize]

Returns the offset values in the offsets buffer.

See Self::offsets for more details.

pub fn value_length(&self, i: usize) -> OffsetSize

Returns the length for value at index i.

pub fn iter<'a>(&'a self) -> ArrayIter<&'a GenericListArray<OffsetSize>>

constructs a new iterator

pub fn slice( &self, offset: usize, length: usize, ) -> GenericListArray<OffsetSize>

Returns a zero-copy slice of this array with the indicated offset and length.

Notes: this method does NOT slice the underlying values array or modify the values in the offsets buffer. See Self::values and Self::offsets for more information.

pub fn from_iter_primitive<T, P, I>(iter: I) -> GenericListArray<OffsetSize>
where T: ArrowPrimitiveType, P: IntoIterator<Item = Option<<T as ArrowPrimitiveType>::Native>>, I: IntoIterator<Item = Option<P>>,

Creates a GenericListArray from an iterator of primitive values

Example

let data = vec![
   Some(vec![Some(0), Some(1), Some(2)]),
   None,
   Some(vec![Some(3), None, Some(5)]),
   Some(vec![Some(6), Some(7)]),
];
let list_array = ListArray::from_iter_primitive::<Int32Type, _, _>(data);
println!("{:?}", list_array);