re_sdk::external::re_types_core::external::arrow::array
Trait BinaryArrayType

pub trait BinaryArrayType<'a>: Sized + ArrayAccessor<Item = &'a [u8]> {
    // Required method
    fn iter(&self) -> ArrayIter<Self> ;
}

A trait for Arrow String Arrays, currently three types are supported:

BinaryArray

LargeBinaryArray

BinaryViewArray

This trait helps to abstract over the different types of binary arrays so that we don’t need to duplicate the implementation for each type.

Required Methods

fn iter(&self) -> ArrayIter<Self>

Constructs a new iterator

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.