re_sdk::external::re_types_core::external::arrow::buffer
Struct NullBuffer

pub struct NullBuffer {
    buffer: BooleanBuffer,
    null_count: usize,
}

A BooleanBuffer used to encode validity for Arrow arrays

In the Arrow specification, array validity is encoded in a packed bitmask with a true value indicating the corresponding slot is not null, and false indicating that it is null.

NullBuffers can be creating using NullBufferBuilder

Fields
buffer: BooleanBuffernull_count: usize
Implementations

impl NullBuffer

pub fn new(buffer: BooleanBuffer) -> NullBuffer

Create a new NullBuffer computing the null count

pub fn new_null(len: usize) -> NullBuffer

Create a new NullBuffer of length len where all values are null

pub fn new_valid(len: usize) -> NullBuffer

Create a new NullBuffer of length len where all values are valid

Note: it is more efficient to not set the null buffer if it is known to be all valid (aka all values are not null)

pub unsafe fn new_unchecked( buffer: BooleanBuffer, null_count: usize, ) -> NullBuffer

Create a new NullBuffer with the provided buffer and null_count

Safety

buffer must contain null_count 0 bits

pub fn union( lhs: Option<&NullBuffer>, rhs: Option<&NullBuffer>, ) -> Option<NullBuffer>

Computes the union of the nulls in two optional NullBuffer

This is commonly used by binary operations where the result is NULL if either of the input values is NULL. Handling the null mask separately in this way can yield significant performance improvements over an iterator approach

pub fn contains(&self, other: &NullBuffer) -> bool

Returns true if all nulls in other also exist in self

pub fn expand(&self, count: usize) -> NullBuffer

Returns a new NullBuffer where each bit in the current null buffer is repeated count times. This is useful for masking the nulls of the child of a FixedSizeListArray based on its parent

pub fn len(&self) -> usize

Returns the length of this NullBuffer

pub fn offset(&self) -> usize

Returns the offset of this NullBuffer in bits

pub fn is_empty(&self) -> bool

Returns true if this NullBuffer is empty

pub fn shrink_to_fit(&mut self)

Free up unused memory.

pub fn null_count(&self) -> usize

Returns the null count for this NullBuffer

pub fn is_valid(&self, idx: usize) -> bool

Returns true if the value at idx is not null

pub fn is_null(&self, idx: usize) -> bool

Returns true if the value at idx is null

pub fn validity(&self) -> &[u8]

Returns the packed validity of this NullBuffer not including any offset

pub fn slice(&self, offset: usize, len: usize) -> NullBuffer

Slices this NullBuffer by the provided offset and length

pub fn iter(&self) -> BitIterator<'_>

Returns an iterator over the bits in this NullBuffer

true indicates that the corresponding value is not NULL

false indicates that the corresponding value is NULL

Note: Self::valid_indices will be significantly faster for most use-cases

pub fn valid_indices(&self) -> BitIndexIterator<'_>

Returns a BitIndexIterator over the valid indices in this NullBuffer

Valid indices indicate the corresponding value is not NULL

pub fn valid_slices(&self) -> BitSliceIterator<'_>

Returns a BitSliceIterator yielding contiguous ranges of valid indices

Valid indices indicate the corresponding value is not NULL

pub fn try_for_each_valid_idx<E, F>(&self, f: F) -> Result<(), E>
where F: FnMut(usize) -> Result<(), E>,

Calls the provided closure for each index in this null mask that is set

pub fn inner(&self) -> &BooleanBuffer

Returns the inner BooleanBuffer

pub fn into_inner(self) -> BooleanBuffer

Returns the inner BooleanBuffer

pub fn buffer(&self) -> &Buffer

Returns the underlying Buffer