Module re_sdk::external::re_types_core::external::arrow::compute::kernels::aggregate

Defines aggregations over Arrow arrays.

Functions

bit_and

Returns the bitwise and of all non-null input values.

bit_or

Returns the bitwise or of all non-null input values.

bit_xor

Returns the bitwise xor of all non-null input values.

bool_and

Returns true if all non-null input values are true, otherwise false.

bool_or

Returns true if any non-null input value is true, otherwise false.

max

Returns the maximum value in the array, according to the natural order. For floating point arrays any NaN values are considered to be greater than any other non-null value

max_array

Returns the max of values in the array of ArrowNumericType type, or dictionary array with value of ArrowNumericType type.

max_binary

Returns the maximum value in the binary array, according to the natural order.

max_binary_view

Returns the maximum value in the binary view array, according to the natural order.

max_boolean

Returns the maximum value in the boolean array

max_string

Returns the maximum value in the string array, according to the natural order.

max_string_view

Returns the maximum value in the string view array, according to the natural order.

min

Returns the minimum value in the array, according to the natural order. For floating point arrays any NaN values are considered to be greater than any other non-null value

min_array

Returns the min of values in the array of ArrowNumericType type, or dictionary array with value of ArrowNumericType type.

min_binary

Returns the minimum value in the binary array, according to the natural order.

min_binary_view

Returns the minimum value in the binary view array, according to the natural order.

min_boolean

Returns the minimum value in the boolean array.

min_string

Returns the minimum value in the string array, according to the natural order.

min_string_view

Returns the minimum value in the string view array, according to the natural order.

sum

Returns the sum of values in the primitive array.

sum_array

Returns the sum of values in the array.

sum_array_checked

Returns the sum of values in the array.

sum_checked

Returns the sum of values in the primitive array.