re_sdk::external::re_types_core::external::arrow::compute::kernels::boolean
Function or

pub fn or(
    left: &BooleanArray,
    right: &BooleanArray,
) -> Result<BooleanArray, ArrowError>

Performs OR operation on two arrays. If either left or right value is null then the result is also null.

Error

This function errors when the arrays have different lengths.

Example

let a = BooleanArray::from(vec![Some(false), Some(true), None]);
let b = BooleanArray::from(vec![Some(true), Some(true), Some(false)]);
let or_ab = or(&a, &b).unwrap();
assert_eq!(or_ab, BooleanArray::from(vec![Some(true), Some(true), None]));