Module re_sdk::external::re_types_core::external::arrow::ipc::writer

Arrow IPC File and Stream Writers

The FileWriter and StreamWriter have similar interfaces, however the FileWriter expects a reader that supports Seeking

Structs

DictionaryTracker

Keeps track of dictionaries that have been written, to avoid emitting the same dictionary multiple times.

EncodedData

Stores the encoded data, which is an crate::Message, and optional Arrow data

FileWriter

Writer for an IPC file

IpcDataGenerator

Handles low level details of encoding Array and Schema into the Arrow IPC Format.

IpcWriteOptions

IPC write options used to control the behaviour of the IpcDataGenerator

StreamWriter

Writer for an IPC stream

Functions

write_message

Write a message’s IPC data and buffers, returning metadata and buffer data lengths written