re_sdk
Module log

Things directly related to logging.

Structs

Chunk

Dense arrow-based storage of N rows of multi-component multi-temporal data for a specific entity.

ChunkBatcher

Implements an asynchronous batcher that coalesces PendingRows into Chunks based upon the thresholds defined in the associated ChunkBatcherConfig.

ChunkBatcherConfig

Defines the different thresholds of the associated ChunkBatcher.

ChunkComponents

ChunkId

A unique ID for a Chunk.

PendingRow

A single row’s worth of data (i.e. a single log call).

RowId

A unique ID for a row’s worth of data within a chunk.

TimeColumn

Enums

ChunkBatcherError

Errors that can occur when creating/manipulating a ChunkBatcher.

ChunkError

Errors that can occur when creating/manipulating a Chunks, directly or indirectly through the use of a crate::ChunkBatcher.

LogMsg

The most general log message sent from the SDK to the server.

Type Aliases

ChunkBatcherResult

ChunkResult