re_sdk
Module log_sink

Structs

BufferedSink

Store log messages in memory until you call LogSink::drain_backlog.

CallbackSink

A sink which forwards all log messages to a callback without any buffering.

GrpcSink

Stream log messages to an a remote Rerun server.

MemorySink

Store log messages directly in memory.

MemorySinkStorage

The storage used by MemorySink.

MemorySinkStorageInner 🔒

Traits

LogSink

Where the SDK sends its log messages.

Type Aliases

LogMsgCallback 🔒