Trait re_sdk::log_sink::LogSink

pub trait LogSink: Send + Sync + 'static {
    fn send(&self, msg: LogMsg);

    fn send_all(&self, messages: Vec<LogMsg>) { ... }
    fn drain_backlog(&self) -> Vec<LogMsg>  { ... }
    fn flush(&self) { ... }
    fn drop_msgs_if_disconnected(&self) { ... }
    fn is_enabled(&self) -> bool { ... }
}

Where the SDK sends its log messages.

Required Methods

fn send(&self, msg: LogMsg)

Send this log message.

Provided Methods

fn send_all(&self, messages: Vec<LogMsg>)

Send all these log messages.

fn drain_backlog(&self) -> Vec<LogMsg>

Drain all buffered LogMsges and return them.

fn flush(&self)

Wait until all logged data have been sent to the remove server (if any).

fn drop_msgs_if_disconnected(&self)

If the TCP session is disconnected, allow it to quit early and drop unsent messages.

fn is_enabled(&self) -> bool

Returns false if this sink just discards all messages.