Module re_sdk::prelude

Structs

AnnotationContext

Archetype: The AnnotationContext provides additional information on how to display entities.

Arrows3D

Archetype: A batch of 3D arrows with optional colors, radii, labels, etc.

Asset3D

Archetype: A prepacked 3D asset (.gltf, .glb, .obj, etc.).

BarChart

Archetype: A bar chart.

Boxes2D

Archetype: A batch of 2d boxes with half-extents and optional center, rotations, rotations, colors etc.

Boxes3D

Archetype: A batch of 3d boxes with half-extents and optional center, rotations, rotations, colors etc.

ClassDescription

Datatype: The description of a semantic Class.

Clear

Archetype: Empties all the components of an entity.

Color

Component: An RGBA color with unmultiplied/separate alpha, in sRGB gamma space with linear alpha.

DepthImage

Archetype: A depth image.

DisconnectedSpace

Archetype: Specifies that the entity path at which this is logged is disconnected from its parent.

Float32

Datatype: A single-precision 32-bit IEEE 754 floating point number.

HalfSizes2D

Component: Half-sizes (extents) of a 2D box along its local axis, starting at its local origin/center.

HalfSizes3D

Component: Half-sizes (extents) of a 3D box along its local axis, starting at its local origin/center.

Image

Archetype: A monochrome or color image.

InstanceKey

Component: A unique numeric identifier for each individual instance within a batch.

KeypointPair

Datatype: A connection between two Keypoints.

LineStrips2D

Archetype: A batch of line strips with positions and optional colors, radii, labels, etc.

LineStrips3D

Archetype: A batch of line strips with positions and optional colors, radii, labels, etc.

Mat3x3

Datatype: A 3x3 Matrix.

Material

Component: Material properties of a mesh.

MediaType

Component: A standardized media type (RFC2046, formerly known as MIME types), encoded as a utf8 string.

Mesh3D

Archetype: A 3D triangle mesh as specified by its per-mesh and per-vertex properties.

MeshProperties

Component: Optional triangle indices for a mesh.

OutOfTreeTransform3D

Component: An out-of-tree affine transform between two 3D spaces, represented in a given direction.

Pinhole

Archetype: Camera perspective projection (a.k.a. intrinsics).

Points2D

Archetype: A 2D point cloud with positions and optional colors, radii, labels, etc.

Points3D

Archetype: A 3D point cloud with positions and optional colors, radii, labels, etc.

Position3D

Component: A position in 3D space.

Quaternion

Datatype: A Quaternion represented by 4 real numbers.

Radius

Component: A Radius component.

Rgba32

Datatype: An RGBA color with unmultiplied/separate alpha, in sRGB gamma space with linear alpha.

RotationAxisAngle

Datatype: 3D rotation represented by a rotation around a given axis.

SegmentationImage

Archetype: An image made up of integer class-ids

Tensor

Archetype: A generic n-dimensional Tensor.

TextDocument

Archetype: A text element intended to be displayed in its own text-box.

TextLog

Archetype: A log entry in a text log, comprised of a text body and its log level.

TextLogLevel

Component: The severity level of a text log message.

Time

A date-time represented as nanoseconds since unix epoch

TimePoint

A point in time on any number of Timelines.

TimeSeriesScalar

Archetype: Log a double-precision scalar that will be visualized as a time-series plot.

Timeline

A time frame/space, e.g. log_time or frame_nr, coupled with the type of time it keeps.

Transform3D

Archetype: A 3D transform.

TranslationAndMat3x3

Datatype: Representation of an affine transform via a 3x3 affine matrix paired with a translation.

TranslationRotationScale3D

Datatype: Representation of an affine transform via separate translation, rotation & scale.

Vector3D

Component: A vector in 3D space.

ViewCoordinates

Archetype: How we interpret the coordinate system of an entity/space.

Enums

Angle

Datatype: Angle in either radians or degrees.

Rotation3D

Datatype: A 3D rotation.

Scale3D

Datatype: 3D scaling factor, part of a transform representation.