Struct re_sdk::ViewCoordinates

#[repr(transparent)]
pub struct ViewCoordinates {
    pub xyz: ViewCoordinates,
}

Archetype: How we interpret the coordinate system of an entity/space.

For instance: What is “up”? What does the Z axis mean? Is this right-handed or left-handed?

The three coordinates are always ordered as [x, y, z].

For example [Right, Down, Forward] means that the X axis points to the right, the Y axis points down, and the Z axis points forward.

Example

view_coordinates_simple:

//! Change the view coordinates for the scene.

fn main() -> Result<(), Box<dyn std::error::Error>> {
    let (rec, storage) =
        rerun::RecordingStreamBuilder::new("rerun_example_view_coordinates").memory()?;

    rec.log_timeless("world", &rerun::ViewCoordinates::RIGHT_HAND_Z_UP)?; // Set an up-axis
    rec.log(
        "world/xyz",
        &rerun::Arrows3D::from_vectors(
            [[1.0, 0.0, 0.0], [0.0, 1.0, 0.0], [0.0, 0.0, 1.0]], //
        )
        .with_colors([[255, 0, 0], [0, 255, 0], [0, 0, 255]]),
    )?;

    rerun::native_viewer::show(storage.take())?;
    Ok(())
}

Fields
xyz: ViewCoordinates
Implementations

impl ViewCoordinates

pub const NUM_COMPONENTS: usize = 3usize

impl ViewCoordinates

pub fn new(xyz: impl Into<ViewCoordinates>) -> ViewCoordinates

impl ViewCoordinates

pub const ULF: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Up, ViewDir::Left, ViewDir::Forward),}

pub const UFL: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Up, ViewDir::Forward, ViewDir::Left),}

pub const LUF: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Left, ViewDir::Up, ViewDir::Forward),}

pub const LFU: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Left, ViewDir::Forward, ViewDir::Up),}

pub const FUL: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Forward, ViewDir::Up, ViewDir::Left),}

pub const FLU: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Forward, ViewDir::Left, ViewDir::Up),}

pub const ULB: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Up, ViewDir::Left, ViewDir::Back),}

pub const UBL: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Up, ViewDir::Back, ViewDir::Left),}

pub const LUB: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Left, ViewDir::Up, ViewDir::Back),}

pub const LBU: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Left, ViewDir::Back, ViewDir::Up),}

pub const BUL: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Back, ViewDir::Up, ViewDir::Left),}

pub const BLU: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Back, ViewDir::Left, ViewDir::Up),}

pub const URF: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Up, ViewDir::Right, ViewDir::Forward),}

pub const UFR: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Up, ViewDir::Forward, ViewDir::Right),}

pub const RUF: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Up, ViewDir::Forward),}

pub const RFU: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Forward, ViewDir::Up),}

pub const FUR: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Forward, ViewDir::Up, ViewDir::Right),}

pub const FRU: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Forward, ViewDir::Right, ViewDir::Up),}

pub const URB: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Up, ViewDir::Right, ViewDir::Back),}

pub const UBR: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Up, ViewDir::Back, ViewDir::Right),}

pub const RUB: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Up, ViewDir::Back),}

pub const RBU: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Back, ViewDir::Up),}

pub const BUR: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Back, ViewDir::Up, ViewDir::Right),}

pub const BRU: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Back, ViewDir::Right, ViewDir::Up),}

pub const DLF: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Down, ViewDir::Left, ViewDir::Forward),}

pub const DFL: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Down, ViewDir::Forward, ViewDir::Left),}

pub const LDF: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Left, ViewDir::Down, ViewDir::Forward),}

pub const LFD: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Left, ViewDir::Forward, ViewDir::Down),}

pub const FDL: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Forward, ViewDir::Down, ViewDir::Left),}

pub const FLD: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Forward, ViewDir::Left, ViewDir::Down),}

pub const DLB: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Down, ViewDir::Left, ViewDir::Back),}

pub const DBL: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Down, ViewDir::Back, ViewDir::Left),}

pub const LDB: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Left, ViewDir::Down, ViewDir::Back),}

pub const LBD: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Left, ViewDir::Back, ViewDir::Down),}

pub const BDL: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Back, ViewDir::Down, ViewDir::Left),}

pub const BLD: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Back, ViewDir::Left, ViewDir::Down),}

pub const DRF: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Down, ViewDir::Right, ViewDir::Forward),}

pub const DFR: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Down, ViewDir::Forward, ViewDir::Right),}

pub const RDF: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Down, ViewDir::Forward),}

pub const RFD: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Forward, ViewDir::Down),}

pub const FDR: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Forward, ViewDir::Down, ViewDir::Right),}

pub const FRD: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Forward, ViewDir::Right, ViewDir::Down),}

pub const DRB: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Down, ViewDir::Right, ViewDir::Back),}

pub const DBR: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Down, ViewDir::Back, ViewDir::Right),}

pub const RDB: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Down, ViewDir::Back),}

pub const RBD: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Back, ViewDir::Down),}

pub const BDR: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Back, ViewDir::Down, ViewDir::Right),}

pub const BRD: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Back, ViewDir::Right, ViewDir::Down),}

pub const RIGHT_HAND_X_UP: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Up, ViewDir::Right, ViewDir::Forward),}

pub const RIGHT_HAND_X_DOWN: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Down, ViewDir::Right, ViewDir::Back),}

pub const RIGHT_HAND_Y_UP: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Up, ViewDir::Back),}

pub const RIGHT_HAND_Y_DOWN: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Down, ViewDir::Forward),}

pub const RIGHT_HAND_Z_UP: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Forward, ViewDir::Up),}

pub const RIGHT_HAND_Z_DOWN: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Back, ViewDir::Down),}

pub const LEFT_HAND_X_UP: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Up, ViewDir::Right, ViewDir::Back),}

pub const LEFT_HAND_X_DOWN: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Down, ViewDir::Right, ViewDir::Forward),}

pub const LEFT_HAND_Y_UP: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Up, ViewDir::Forward),}

pub const LEFT_HAND_Y_DOWN: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Down, ViewDir::Back),}

pub const LEFT_HAND_Z_UP: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Back, ViewDir::Up),}

pub const LEFT_HAND_Z_DOWN: ViewCoordinates = Self{ xyz: components::ViewCoordinates::new(ViewDir::Right, ViewDir::Forward, ViewDir::Down),}