Crate re_sdk

The Rerun SDK

Feature flags

demo (enabled by default) — Enable the demo module (helpers for Rerun examples).

glam (enabled by default) — Add support for some math operations using glam.

image (enabled by default) — Integration with the image crate, plus JPEG support..

web_viewer — Support serving a web viewer over HTTP.

Enabling this inflates the binary size quite a bit, since it embeds the viewer wasm.

For faster & easier builds, the web-player build comes bundled with the crate. Enabling this feature will embed this pre-built web-player.

However, when building from source in the repository, this feature adds quite a bit to the compile time since it requires compiling and bundling the viewer as wasm. You also need to install some additional tools, which you can do by running scripts/setup_web.sh.

Modules

archetypes

components

These are the different components you can log.

coordinates

Coordinate system helpers, for use with components::ViewCoordinates.

datatypes

demo_util

This module contains utilities to support Rerun examples.

external

Re-exports of other crates.

global 🔒

Keeps track of global and thread-local RecordingStreams and handles fallback logic between them.

log

Things directly related to logging.

log_sink 🔒

msg_sender 🔒

recording_stream 🔒

sink

Different destinations for log messages.

time

Time-related types.

transform

Transform helpers, for use with components::Transform3D.

web_viewer

Methods for spawning the web viewer and streaming the SDK log stream to it.

Structs

ApplicationId

The user-chosen name of the application doing the logging.

ComponentName

The name of an entity component, e.g. pos or color.

EntityPath

camera / "left" / points / #42

MsgSender

Facilitates building and sending component payloads with the Rerun SDK.

RecordingStream

A RecordingStream handles everything related to logging data into Rerun.

RecordingStreamBuilder

Construct a RecordingStream.

StoreId

A unique id per store.

Enums

MsgSenderError

Errors that can occur when constructing or sending messages using MsgSender.

StoreKind

What kind of Store this is.

Constants

RERUN_ENV_VAR 🔒

Traits

Archetype

Component

Datatype

A Datatype describes plain old data that can be used by any number of Component.

LegacyComponent

A type that can used as a Component of an Entity.

Loggable

Anything that can be serialized to and deserialized from Arrow data.

Functions

build_info

The version of the Rerun SDK.

called_from_official_rust_example 🔒

cleanup_if_forked_child

Check whether we are the child of a fork.

decide_logging_enabled

Checks the RERUN environment variable. If not found, returns the argument.

default_flush_timeout

The default amount of time to wait for the TCP connection to resume during a flush

default_server_addr

The default address of a Rerun TCP server which an SDK connects to.

get_rerun_env 🔒

Helper to get the value of the RERUN environment variable.

new_store_info

Creates a new re_log_types::StoreInfo which can be used with RecordingStream::new.