Crate re_sdk

The Rerun SDK

Feature flags

demo (enabled by default) — Enable the demo module (helpers for Rerun examples).

glam (enabled by default) — Add support for some math operations using glam.

global_session — Add the global_session method.

image (enabled by default) — Integration with the image crate.

Modules

components

These are the different components you can log.

coordinates

Coordinate system helpers, for use with components::ViewCoordinates.

demo_util

This module contains utilities to support Rerun examples.

external

Re-exports of other crates.

file_sink 🔒

global 🔒

log

Things directly related to logging.

log_sink 🔒

msg_sender 🔒

session 🔒

sink

Different destinations for log messages.

time

Time-related types.

Structs

ApplicationId

The user-chosen name of the application doing the logging.

ComponentName

The name of an entity component, e.g. pos or color.

EntityPath

camera / "left" / points / #42

MsgSender

Facilitates building and sending component payloads with the Rerun SDK.

RecordingId

A unique id per recording (a stream of LogMsges).

Session

The main way to do Rerun loggning.

SessionBuilder

Construct a Session.

Enums

MsgSenderError

Errors that can occur when constructing or sending messages using MsgSender.

Constants

RERUN_ENV_VAR 🔒

Traits

Component

A type that can used as a Component of an Entity.

SerializableComponent

A Component that fulfils all the conditions required to be serialized as an Arrow payload.

Functions

called_from_official_rust_example 🔒

decide_logging_enabled

Checks the RERUN environment variable. If not found, returns the argument.

default_server_addr

The default address of a Rerun TCP server which an SDK connects to.

get_rerun_env 🔒

Helper to get the value of the RERUN environment variable.

global_session

Access a global Session singleton for convenient logging.

new_recording_info

Creates a new [re_log_types::RecordingInfo] which can be used with Session::new.