Crate re_sdk

The Rerun logging SDK

This is the bare-bones version of the rerun crate. rerun exports everything in re_sdk, so in most cases you want to use rerun instead.

Please read the docs for the rerun crate instead.

Feature flags

demo (enabled by default) — Enable the demo module (helpers for Rerun examples).

glam (enabled by default) — Add support for some math operations using glam.

image (enabled by default) — Integration with the image crate, plus JPEG support.

log (enabled by default) — Integration with the log crate.

web_viewer — Support serving a web viewer over HTTP.

Enabling this inflates the binary size quite a bit, since it embeds the viewer wasm.

For faster & easier builds, a pre-built web-views comes bundled with the crate as a Wasm binary. Enabling this feature will embed this pre-built web viewer.

However, when building from source in the repository, this feature adds quite a bit to the compile time since it requires compiling and bundling the viewer as wasm. You also need to install some additional tools, which you can do by running scripts/setup_web.sh.

Modules

archetypes

Acrchetype are the high-level things you can log, like Image, Points3D, etc.

components

Components are the basic building blocks of archetypes.

coordinates

Coordinate system helpers, for use with components::ViewCoordinates.

datatypes

The low-level datatypes that components are built from.

demo_util

This module contains utilities to support Rerun examples.

external

Re-exports of other crates.

global 🔒

Keeps track of global and thread-local RecordingStreams and handles fallback logic between them.

log

Things directly related to logging.

log_integration 🔒

log_sink 🔒

prelude 🔒

recording_stream 🔒

sink

Different destinations for log messages.

time

Time-related types.

transform

Transform helpers, for use with components::Transform3D.

web_viewer

Methods for spawning the web viewer and streaming the SDK log stream to it.

Structs

AnnotationContext

Archetype: The AnnotationContext provides additional information on how to display entities.

ApplicationId

The user-chosen name of the application doing the logging.

ArchetypeName

The fully-qualified name of an Archetype, e.g. rerun.archetypes.Points3D.

Arrows3D

Archetype: A batch of 3D arrows with optional colors, radii, labels, etc.

Asset3D

Archetype: A prepacked 3D asset (.gltf, .glb, .obj, etc.).

BarChart

Archetype: A bar chart.

Boxes2D

Archetype: A batch of 2d boxes with half-extents and optional center, rotations, rotations, colors etc.

Boxes3D

Archetype: A batch of 3d boxes with half-extents and optional center, rotations, rotations, colors etc.

ClassDescription

Datatype: The description of a semantic Class.

Clear

Archetype: Empties all the components of an entity.

Color

Component: An RGBA color with unmultiplied/separate alpha, in sRGB gamma space with linear alpha.

ComponentName

The fully-qualified name of a Component, e.g. rerun.components.Position2D.

DatatypeName

The fully-qualified name of a Datatype, e.g. rerun.datatypes.Vec2D.

DepthImage

Archetype: A depth image.

DisconnectedSpace

Archetype: Specifies that the entity path at which this is logged is disconnected from its parent.

EntityPath

The unique identifier of an entity, e.g. camera/"ACME Örnöga"/points

Float32

Datatype: A single-precision 32-bit IEEE 754 floating point number.

GenericIndicatorComponent

A generic indicator component that can be specialized for any Archetype.

HalfSizes2D

Component: Half-sizes (extents) of a 2D box along its local axis, starting at its local origin/center.

HalfSizes3D

Component: Half-sizes (extents) of a 3D box along its local axis, starting at its local origin/center.

Image

Archetype: A monochrome or color image.

InstanceKey

Component: A unique numeric identifier for each individual instance within a batch.

KeypointPair

Datatype: A connection between two Keypoints.

LineStrips2D

Archetype: A batch of line strips with positions and optional colors, radii, labels, etc.

LineStrips3D

Archetype: A batch of line strips with positions and optional colors, radii, labels, etc.

Logger

Implements a log::Log that forwards all events to the Rerun SDK.

Mat3x3

Datatype: A 3x3 Matrix.

Material

Component: Material properties of a mesh.

MediaType

Component: A standardized media type (RFC2046, formerly known as MIME types), encoded as a utf8 string.

Mesh3D

Archetype: A 3D triangle mesh as specified by its per-mesh and per-vertex properties.

MeshProperties

Component: Optional triangle indices for a mesh.

NamedIndicatorComponent

An arbitrary named indicator component.

OutOfTreeTransform3D

Component: An out-of-tree affine transform between two 3D spaces, represented in a given direction.

Pinhole

Archetype: Camera perspective projection (a.k.a. intrinsics).

Points2D

Archetype: A 2D point cloud with positions and optional colors, radii, labels, etc.

Points3D

Archetype: A 3D point cloud with positions and optional colors, radii, labels, etc.

Position3D

Component: A position in 3D space.

Quaternion

Datatype: A Quaternion represented by 4 real numbers.

Radius

Component: A Radius component.

RecordingStream

A RecordingStream handles everything related to logging data into Rerun.

RecordingStreamBuilder

Construct a RecordingStream.

Rgba32

Datatype: An RGBA color with unmultiplied/separate alpha, in sRGB gamma space with linear alpha.

RotationAxisAngle

Datatype: 3D rotation represented by a rotation around a given axis.

SegmentationImage

Archetype: An image made up of integer class-ids

StoreId

A unique id per store.

Tensor

Archetype: A generic n-dimensional Tensor.

TextDocument

Archetype: A text element intended to be displayed in its own text-box.

TextLog

Archetype: A log entry in a text log, comprised of a text body and its log level.

TextLogLevel

Component: The severity level of a text log message.

Time

A date-time represented as nanoseconds since unix epoch

TimePoint

A point in time on any number of Timelines.

TimeSeriesScalar

Archetype: Log a double-precision scalar that will be visualized as a time-series plot.

Timeline

A time frame/space, e.g. log_time or frame_nr, coupled with the type of time it keeps.

Transform3D

Archetype: A 3D transform.

TranslationAndMat3x3

Datatype: Representation of an affine transform via a 3x3 affine matrix paired with a translation.

TranslationRotationScale3D

Datatype: Representation of an affine transform via separate translation, rotation & scale.

Vector3D

Component: A vector in 3D space.

ViewCoordinates

Archetype: How we interpret the coordinate system of an entity/space.

Enums

Angle

Datatype: Angle in either radians or degrees.

MaybeOwnedComponentBatch

Holds either an owned ComponentBatch that lives on heap, or a reference to one.

Rotation3D

Datatype: A 3D rotation.

Scale3D

Datatype: 3D scaling factor, part of a transform representation.

StoreKind

What kind of Store this is.

Constants

RERUN_ENV_VAR 🔒

Traits

Archetype

An archetype is a high-level construct that represents a set of Components that usually play well with each other (i.e. they compose nicely).

AsComponents

Describes the interface for interpreting an object as a bundle of Components.

Component

A Component describes semantic data that can be used by any number of Archetypes.

ComponentBatch

A ComponentBatch represents an array’s worth of Component instances.

Datatype

A Datatype describes plain old data that can be used by any number of Components.

DatatypeBatch

A DatatypeBatch represents an array’s worth of Datatype instances.

Loggable

A Loggable represents a single instance in an array of loggable data.

LoggableBatch

A LoggableBatch represents an array’s worth of Loggable instances, ready to be serialized.

Functions

build_info

The version of the Rerun SDK.

called_from_official_rust_example 🔒

cleanup_if_forked_child

Check whether we are the child of a fork.

decide_logging_enabled

Checks the RERUN environment variable. If not found, returns the argument.

default_flush_timeout

The default amount of time to wait for the TCP connection to resume during a flush

default_log_filter

Get RUST_LOG environment variable or info, if not set.

default_server_addr

The default address of a Rerun TCP server which an SDK connects to.

get_rerun_env 🔒

Helper to get the value of the RERUN environment variable.

new_store_info

Creates a new re_log_types::StoreInfo which can be used with RecordingStream::new.

Type Definitions

RecordingStreamResult

Results that can occur when creating/manipulating a RecordingStream.