Crate re_sdk

The Rerun logging SDK

This is the bare-bones version of the rerun crate. rerun exports everything in re_sdk, so in most cases you want to use rerun instead.

Please read the docs for the rerun crate instead.

Feature flags

data_loaders — Support for using Rerun’s data-loaders directly from the SDK.

See our log_file example and https://www.rerun.io/docs/reference/data-loaders/overview for more information.

web_viewer — Support serving a web viewer over HTTP.

Enabling this inflates the binary size quite a bit, since it embeds the viewer wasm.

For faster & easier builds, a pre-built web-views comes bundled with the crate as a Wasm binary. Enabling this feature will embed this pre-built web viewer.

However, when building from source in the repository, this feature adds quite a bit to the compile time since it requires compiling and bundling the viewer as wasm.

Modules

binary_stream_sink 🔒

external

Re-exports of other crates.

global 🔒

Keeps track of global and thread-local RecordingStreams and handles fallback logic between them.

log

Things directly related to logging.

log_sink 🔒

recording_stream 🔒

sink

Different destinations for log messages.

spawn 🔒

time

Time-related types.

web_viewer

Methods for spawning the web viewer and streaming the SDK log stream to it.

Macros

entity_path

Build an EntityPath from parts that are not escaped:

Structs

ApplicationId

The user-chosen name of the application doing the logging.

ArchetypeName

The fully-qualified name of an Archetype, e.g. rerun.archetypes.Points3D.

ComponentName

The fully-qualified name of a Component, e.g. rerun.components.Position2D.

DataLoaderSettings

Recommended settings for the DataLoader.

DatatypeName

The fully-qualified name of a Datatype, e.g. rerun.datatypes.Vec2D.

EntityPath

The unique identifier of an entity, e.g. camera/3/points

EntityPathPart

The different parts that make up an EntityPath.

GenericIndicatorComponent

A generic indicator component that can be specialized for any Archetype.

Instance

A unique numeric index for each individual instance within a batch.

MemoryLimit

Represents a limit in how much RAM to use for the entire process.

NamedIndicatorComponent

An arbitrary named indicator component.

RecordingStream

A RecordingStream handles everything related to logging data into Rerun.

RecordingStreamBuilder

Construct a RecordingStream.

SpawnOptions

Options to control the behavior of spawn.

StoreId

A unique id per store.

Time

A date-time represented as nanoseconds since unix epoch

TimePoint

A point in time on any number of Timelines.

Timeline

A time frame/space, e.g. log_time or frame_nr, coupled with the type of time it keeps.

Enums

DataLoaderError

Errors that might happen when loading data through a DataLoader.

LoadedData

What DataLoaders load.

MaybeOwnedComponentBatch

Holds either an owned ComponentBatch that lives on heap, or a reference to one.

RecordingStreamError

Errors that can occur when creating/manipulating a RecordingStream.

SpawnError

Errors that can occur when spawning a Rerun Viewer.

StoreKind

What kind of Store this is.

Constants

RERUN_ENV_VAR 🔒

Traits

Archetype

An archetype is a high-level construct that represents a set of Components that usually play well with each other (i.e. they compose nicely).

AsComponents

Describes the interface for interpreting an object as a bundle of Components.

Component

A Component describes semantic data that can be used by any number of Archetypes.

ComponentBatch

A ComponentBatch represents an array’s worth of Component instances.

DataLoader

A DataLoader loads data from a file path and/or a file’s contents.

Datatype

A Datatype describes plain old data that can be used by any number of Components.

DatatypeBatch

A DatatypeBatch represents an array’s worth of Datatype instances.

Loggable

A Loggable represents a single instance in an array of loggable data.

LoggableBatch

A LoggableBatch represents an array’s worth of Loggable instances, ready to be serialized.

SizeBytes

Approximations of stack and heap size for both internal and external types.

Functions

build_info

The version of the Rerun SDK.

called_from_official_rust_example 🔒

cleanup_if_forked_child

Check whether we are the child of a fork.

decide_logging_enabled

Checks the RERUN environment variable. If not found, returns the argument.

default_flush_timeout

The default amount of time to wait for the TCP connection to resume during a flush

default_server_addr

The default address of a Rerun TCP server which an SDK connects to.

forced_sink_path

Returns path for force sink if private environment variable _RERUN_TEST_FORCE_SAVE is set

get_rerun_env 🔒

Helper to get the value of the RERUN environment variable.

new_store_info

Creates a new re_log_types::StoreInfo which can be used with RecordingStream::new.

spawn

Spawns a new Rerun Viewer process ready to listen for TCP connections.

Type Aliases

RecordingStreamResult

Results that can occur when creating/manipulating a RecordingStream.