Module re_sdk::recording_stream

Structs

RecordingStream

A RecordingStream handles everything related to logging data into Rerun.

RecordingStreamBuilder

Construct a RecordingStream.

RecordingStreamInner 🔒

ThreadInfo 🔒

Thread-local data.

Enums

Command 🔒

RecordingStreamError

Errors that can occur when creating/manipulating a RecordingStream.

Constants

ENV_FORCE_SAVE 🔒

Private environment variable meant for tests.

Functions

forced_sink_path

Returns path for force sink if private environment variable _RERUN_TEST_FORCE_SAVE is set

forwarding_thread 🔒

Type Aliases

RecordingStreamResult

Results that can occur when creating/manipulating a RecordingStream.